recommendation/requirements.txt
scikit-surprise==1.1.3
zenml==0.36.0
pytest==7.2.2